move types to response classes file
