+++ package.json
@@ -24,7 +24,6 @@
   },
   "homepage": "https://github.com/sandboxnu/knowyouroptions#readme",
   "devDependencies": {
-    "@svgr/webpack": "^5.5.0",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
     "eslint": "^7.32.0",
